\name{fact}
\alias{fact}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to calculate the factorial.
}
\description{
Calculates the factorial of a defined number.
}
\usage{
fact(n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
a numeric value for which the value of the factorial to be calculated.
}
}
\details{
n should be an integer.
}
\value{
Returns a numeric value.
}
\references{
Ilk, O. (2011). \emph{R Yazilimina Giris} [Introduction to R Language]. ODTU Yayincilik [METU Press].
}
\author{
Ozlem Ilk
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
\examples{
fact(4)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ factorial}

