\name{OOmisc-package}
\alias{OOmisc-package}
\alias{OOmisc}
\docType{package}
\title{
Ozgur-Ozlem Miscellaneous
}
\description{
Includes miscellaneous functions. The package includes several functions with different goals such as random numer generation, numerical root finding, the calculation of accuracy measures and summary statistics etc.
}
\details{
\tabular{ll}{
Package: \tab OOmisc \cr
Version: \tab 1.2 \cr
Date: \tab 2012-02-03 \cr
License: \tab GPL (>=2) \cr
}
}

