\name{OOmisc-package}
\alias{OOmisc-package}
\alias{OOmisc}
\docType{package}
\title{
Includes miscellaneous functions
}
\description{
The package includes several functions with different goals such as generating random numbers, root finding, the calculation of summary statistics etc.
}
\details{
\tabular{ll}{
 Package: \tab OOmisc\cr
Version: \tab 1.0\cr
Date: \tab 2012-04-09\cr
License: \tab GPL (>=2)\cr
}
}

